change redirect file name to game, display quote data on page for quote mode
