+++ game/guess_that_quote.html
@@ -51,7 +51,9 @@
       </div>
   </div>
 
-  <p class="quote" id="quoteAPI"></p>
+  <p class="quote" id="quoteAPI">
+    
+  </p>
 
   <div class="columns centerColumn">
     <div class="column">
