+++ script.js
@@ -1,6 +1,4 @@
-
-import {guessAnimePage, guessQuotePage, guessMangaPage} from '/redirect.js'
-import axios from 'axios'; 
+import {guessAnimePage, guessQuotePage, guessMangaPage} from '/game.js'
 
 
 
@@ -31,20 +29,8 @@ $(document).ready(function() {
 
 /*
 guessthatquote.html 
-
 */
 
 let requestQuotes = 'https://animechan.vercel.app/api/random';
 
-const getQuote = async () => {
-	try {
-		const response = await axios.get(`https://animechan.vercel.app/api/random`)
-        console.log(`Quote Data: ${response.data.quote}`); 
-        // return response.data.quote
 
-	} catch (err) {
-		console.log(err)
-	}
-}
-
-getQuote();
